+++ src/interfaces/fb_audio_chunk.h
@@ -19,7 +19,7 @@ namespace mozjs
 		~JsFbAudioChunk() override = default;
 
 		static std::unique_ptr<JsFbAudioChunk> CreateNative(JSContext* cx, const audio_chunk_impl& chunk);
-		static size_t GetInternalSize(const audio_chunk_impl& chunk);
+		static uint32_t GetInternalSize(const audio_chunk_impl& chunk);
 
 	public:
 		JS::Value get_Data();
