+++ src/interfaces/fb_playback_queue_item.h
@@ -19,7 +19,7 @@ namespace mozjs
 		~JsFbPlaybackQueueItem() override = default;
 
 		static std::unique_ptr<JsFbPlaybackQueueItem> CreateNative(JSContext* cx, const t_playback_queue_item& playbackQueueItem);
-		static size_t GetInternalSize(const t_playback_queue_item& playbackQueueItem);
+		static uint32_t GetInternalSize(const t_playback_queue_item& playbackQueueItem);
 
 	public:
 		JSObject* get_Handle();
