+++ src/interfaces/fb_tooltip.cpp
@@ -95,7 +95,7 @@ namespace mozjs
 		return std::unique_ptr<JsFbTooltip>(new JsFbTooltip(cx, hParentWnd));
 	}
 
-	size_t JsFbTooltip::GetInternalSize(HWND /*hParentWnd*/)
+	uint32_t JsFbTooltip::GetInternalSize(HWND /*hParentWnd*/)
 	{
 		return sizeof(LOGFONT) + sizeof(TOOLINFO);
 	}
